DROOLS-3438: Unify Cell Dom Element factories
